pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cli-research-agent"
version = "0.1.0"
description = "A ReAct-based CLI research agent powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.11.0",
    "tavily-python>=0.3.0",
    "wikipedia>=1.4.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
research = "cli.main:app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]